fulfilled styleHelper file to expose common styles
