+++ screens/ConfirmScreen.js
@@ -1,5 +1,6 @@
 import React from 'react';
-import { View, Text, StyleSheet, Modal, TouchableOpacity } from 'react-native';
+import { View, Text, Modal, TouchableOpacity } from 'react-native';
+import { styles } from '../components/StyleHelper';
 
 export default function ConfirmScreen({ visible, userInfo, onEdit, onContinue }) {
   return (
@@ -28,46 +29,3 @@ export default function ConfirmScreen({ visible, userInfo, onEdit, onContinue })
     </Modal>
   );
 }
-
-const styles = StyleSheet.create({
-  modalBackground: {
-    flex: 1,
-    backgroundColor: 'rgba(0, 0, 0, 0.5)',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  card: {
-    padding: 20,
-    borderRadius: 10,
-    backgroundColor: 'lightgray',
-    alignItems: 'left',
-    justifyContent: 'center',
-  },
-  infoText: {
-    color: 'blue',
-    fontSize: 18,
-    marginVertical: 5,
-    textAlign: 'left',
-  },
-  buttonContainer: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    marginTop: 20,
-    width: '100%',
-  },
-  button: {
-    flex: 1,
-    marginHorizontal: 10,
-    paddingHorizontal: 10,
-    alignItems: 'center',
-    borderRadius: 5,
-  },
-  gobackButton: {
-    color: 'red',
-    fontSize: 16,
-  },
-  continueButton: {
-    color: 'blue',
-    fontSize: 16,
-  },
-});
